// user-server/front-jin/james-app/app/components/user/model/user-column.ts
export interface UserColumn{
    id?: string;
    username: string;
    password: string;
    name?: string;
    age?: string;
    email?: string;
    address?: string;
    phone?: string;
    asset?: string;
    mbti?: string;
    investmentPropensity?: string;
    job?: string;
    regDate?: string;
    modDate?: string;
}